+++ app/layout.tsx
@@ -1,0 +1,25 @@
+'use client';
+
+import { ThemeProvider } from '@mui/material';
+import { theme } from './styles/theme';
+import Navbar from './components/features/Navbar';
+import './globals.css';
+
+export default function RootLayout({
+  // Layouts must accept a children prop.
+  // This will be populated with nested layouts or pages
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body suppressHydrationWarning={true}>
+        <ThemeProvider theme={theme}>
+          <Navbar />
+          {children}
+        </ThemeProvider>
+      </body>
+    </html>
+  );
+}
