+++ app/lib/mongo/index.ts
@@ -1,0 +1,31 @@
+import { GridFSBucket, MongoClient } from 'mongodb';
+
+const url = process.env.DB_URL;
+
+if (!url) throw new Error('Please add your MongoDB url to .env.local');
+
+let client = new MongoClient(url);
+
+let clientPromise: Promise<MongoClient>;
+
+let globalWithMongo = global as typeof globalThis & {
+  _mongoClientPromise: Promise<MongoClient>;
+};
+
+if (process.env.NODE_ENV === 'development') {
+  // In development mode, use a global variable so that the value
+  // is preserved across module reloads caused by HMR (Hot Module Replacement).
+  if (!globalWithMongo._mongoClientPromise) {
+    client = new MongoClient(url);
+    globalWithMongo._mongoClientPromise = client.connect();
+  }
+  clientPromise = globalWithMongo._mongoClientPromise;
+} else {
+  // In production mode, it's best to not use a global variable.
+  client = new MongoClient(url);
+  clientPromise = client.connect();
+}
+
+clientPromise = globalWithMongo._mongoClientPromise;
+
+export default clientPromise;
