+++ app/lib/mongo/shcema.ts
@@ -1,0 +1,8 @@
+import { Schema } from 'mongoose';
+
+interface ImageSchemaProps {
+  image: Buffer;
+}
+export const imageSchema = new Schema<ImageSchemaProps>({
+  image: Buffer,
+});
