+++ next.config.js
@@ -1,9 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  api: {
-    bodyParser: false,
-  },
+  trailingSlash: true,
 };
 
 module.exports = nextConfig;
