--- src/lib/mongo/shcema.ts
@@ -1,8 +1,0 @@
-import { Schema } from 'mongoose';
-
-interface ImageSchemaProps {
-  image: Buffer;
-}
-export const imageSchema = new Schema<ImageSchemaProps>({
-  image: Buffer,
-});
