--- src/pages/_app.tsx
@@ -1,33 +1,0 @@
-import IndexLayout from '@/layout';
-import '@/styles/globals.css';
-import React from 'react';
-import type { ReactElement, ReactNode } from 'react';
-import type { NextPage } from 'next';
-import type { AppProps } from 'next/app';
-import { ThemeProvider } from '@mui/material';
-import { theme } from '@/styles/theme';
-import Head from 'next/head';
-
-export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
-  getLayout?: (page: ReactElement) => ReactNode;
-};
-
-type AppPropsWithLayout = AppProps & {
-  Component: NextPageWithLayout;
-};
-
-export default function MyApp({ Component, pageProps }: AppPropsWithLayout) {
-  return (
-    <ThemeProvider theme={theme}>
-      <IndexLayout>
-        <Head>
-          <title>Plant Diary</title>
-          <meta name="description" content="Generated by create next app" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <Component {...pageProps} />
-      </IndexLayout>
-    </ThemeProvider>
-  );
-}
