--- src/pages/_document.tsx
@@ -1,18 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document';
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head>
-        <link
-          href="https://fonts.googleapis.com/css2?family=Rajdhani:wght@300;400;600&family=Rampart+One&display=swap"
-          rel="stylesheet"
-        />
-      </Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
