--- src/styles/globals.css
@@ -1,15 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  font-family: Rajdhani;
-}
-
-a {
-  text-decoration: none;
-  color: black;
-}
-
-body {
-  background-color: #f5f5f5;
-}
